refactor: 강의평가 디테일 모달 모듈화
